modified the app navigator
