+++ app/actions/user.js
@@ -7,6 +7,7 @@ import { SET_INITIAL_USER_STATE, SET_USER_DETAILS, CLEAR_USER_DETAILS } from './
 import { bugsnag } from '../utils';
 import { checkAndAddUserSpecies } from '../utils/addUserSpecies';
 import { getAuthenticatedRequest, getRequest, postAuthenticatedRequest } from '../utils/api';
+import AsyncStorage from '@react-native-community/async-storage';
 
 // creates auth0 instance while providing the auth0 domain and auth0 client id
 const auth0 = new Auth0({ domain: Config.AUTH0_DOMAIN, clientId: Config.AUTH0_CLIENT_ID });
@@ -109,9 +110,12 @@ export const auth0Logout = (userDispatch = null) => {
   return new Promise((resolve) => {
     auth0.webAuth
       .clearSession()
-      .then(() => {
+      .then(async () => {
         // deletes the user from DB
         deleteUser();
+
+        // removes [isInitialSyncDone] item from AsyncStorage
+        await AsyncStorage.removeItem('isInitialSyncDone');
 
         if (userDispatch) {
           // clear the user details from the user state
