+++ app/components/App/index.js
@@ -1,196 +1,16 @@
 import MapboxGL from '@react-native-mapbox-gl/maps';
-import { NavigationContainer } from '@react-navigation/native';
-import {
-  createStackNavigator,
-  HeaderStyleInterpolators,
-  TransitionSpecs,
-} from '@react-navigation/stack';
 import React from 'react';
 import Config from 'react-native-config';
 import 'react-native-gesture-handler';
-import {
-  CreatePolygon,
-  DownloadMap,
-  InventoryOverview,
-  Legals,
-  LocateTree,
-  Logs,
-  MainScreen,
-  ManageSpecies,
-  ManageUsers,
-  RegisterSingleTree,
-  RegisterTree,
-  SavedAreas,
-  SelectCoordinates,
-  SelectProject,
-  SelectSpecies,
-  SignUp,
-  SingleTreeOverview,
-  TreeInventory,
-  UploadedInventory,
-} from '../';
-import { auth0Logout, getNewAccessToken, getUserDetailsFromServer } from '../../actions/user';
-import SpecieInfo from '../ManageSpecies/SpecieInfo';
-import InitialLoading from '../InitialLoading';
 import Provider from '../../reducers/provider';
-import { getUserDetails } from '../../repositories/user';
-import { dailyLogUpdateCheck } from '../../utils/logs';
-import updateAndSyncLocalSpecies from '../../utils/updateAndSyncLocalSpecies';
-import { migrateRealm } from '../../repositories/default';
-import dbLog from '../../repositories/logs';
-import { LogTypes } from '../../utils/constants';
-import { checkAndAddUserSpecies } from '../../utils/addUserSpecies';
+import AppNavigator from '../Navigator';
 
 MapboxGL.setAccessToken(Config.MAPBOXGL_ACCCESS_TOKEN);
 
-const Stack = createStackNavigator();
-
-const MyTransition = {
-  gestureDirection: 'vertical',
-  transitionSpec: {
-    open: TransitionSpecs.TransitionIOSSpec,
-    close: TransitionSpecs.TransitionIOSSpec,
-  },
-  headerStyleInterpolator: HeaderStyleInterpolators.forFade,
-  cardStyleInterpolator: ({ current, layouts }) => {
-    return {
-      cardStyle: {
-        transform: [
-          {
-            translateX: current.progress.interpolate({
-              inputRange: [0, 1],
-              outputRange: [layouts.screen.width, 0],
-            }),
-          },
-        ],
-      },
-      overlayStyle: {
-        opacity: current.progress.interpolate({
-          inputRange: [0, 1],
-          outputRange: [0, 0.5],
-        }),
-      },
-    };
-  },
-};
-
 const App = () => {
-  const [isDBMigrating, setIsDBMigrating] = React.useState(true);
-  const [areSpeciesLoading, setAreSpeciesLoading] = React.useState(false);
-
-  const checkIsUserLogin = async () => {
-    const dbUserDetails = await getUserDetails();
-
-    if (dbUserDetails && dbUserDetails.refreshToken) {
-      const newAccessToken = await getNewAccessToken(dbUserDetails.refreshToken);
-      if (newAccessToken) {
-        // fetches the user details from server by passing the accessToken which is used while requesting the API
-        getUserDetailsFromServer(newAccessToken);
-        checkAndAddUserSpecies();
-      } else {
-        auth0Logout();
-      }
-    }
-  };
-
-  React.useEffect(() => {
-    // calls the migration function to migrate the realm with [setIsDBMigrating]
-    // as callback param which changes the [isDBMigrating] state.
-    migrateRealm(setIsDBMigrating)
-      .then(() => {
-        // logs success in DB
-        dbLog.info({
-          logType: LogTypes.OTHER,
-          message: 'DB migration successfully done',
-        });
-        // calls this function to update the species in the realm DB with [setAreSpeciesLoading]
-        // as callback param which changes the [areSpeciesLoading] state
-        updateAndSyncLocalSpecies((isLoading) => {
-          // sets [isDBMigrating = false] to hide the MigratingDB screen
-          setAreSpeciesLoading(isLoading);
-          setIsDBMigrating(false);
-        })
-          .then(postUpdatingSpecies)
-          .catch(postUpdatingSpecies);
-      })
-      .catch((err) => {
-        // sets [isDBMigrating = false] to hide the MigratingDB screen
-        setIsDBMigrating(false);
-        console.error(`Error while setting up realm connection - App, ${JSON.stringify(err)}`);
-      });
-  }, []);
-
-  // called after local species are updated
-  const postUpdatingSpecies = () => {
-    // sets [isDBMigrating = false] to hide the MigratingDB screen
-    setIsDBMigrating(false);
-    // sets [areSpeciesLoading = false] to hide the SpeciesLoading screen
-    setAreSpeciesLoading(false);
-    checkIsUserLogin();
-    dailyLogUpdateCheck();
-  };
-
-  let initialRouteName = 'MainScreen';
-  if (isDBMigrating) {
-    initialRouteName = 'MigratingDB';
-  } else if (areSpeciesLoading) {
-    initialRouteName = 'SpeciesLoading';
-  }
-
   return (
     <Provider>
-      <NavigationContainer>
-        <Stack.Navigator initialRouteName={initialRouteName} headerMode={'none'}>
-          {isDBMigrating ? (
-            <Stack.Screen name="MigratingDB" component={InitialLoading} />
-          ) : areSpeciesLoading ? (
-            <Stack.Screen name="SpeciesLoading" component={InitialLoading} />
-          ) : (
-            <>
-              <Stack.Screen name="MainScreen" component={MainScreen} options={MyTransition} />
-              <Stack.Screen name="TreeInventory" component={TreeInventory} options={MyTransition} />
-              <Stack.Screen name="RegisterTree" component={RegisterTree} options={MyTransition} />
-              <Stack.Screen name="SelectProject" component={SelectProject} options={MyTransition} />
-              <Stack.Screen name="LocateTree" component={LocateTree} options={MyTransition} />
-              <Stack.Screen name="CreatePolygon" component={CreatePolygon} options={MyTransition} />
-              <Stack.Screen
-                name="InventoryOverview"
-                component={InventoryOverview}
-                options={MyTransition}
-              />
-              <Stack.Screen name="SavedAreas" component={SavedAreas} options={MyTransition} />
-              <Stack.Screen name="DownloadMap" component={DownloadMap} options={MyTransition} />
-              <Stack.Screen
-                name="RegisterSingleTree"
-                component={RegisterSingleTree}
-                options={MyTransition}
-              />
-              <Stack.Screen
-                name="SingleTreeOverview"
-                component={SingleTreeOverview}
-                options={MyTransition}
-              />
-              <Stack.Screen
-                name="SelectCoordinates"
-                component={SelectCoordinates}
-                options={MyTransition}
-              />
-              <Stack.Screen name="ManageUsers" component={ManageUsers} options={MyTransition} />
-              <Stack.Screen name="SignUp" component={SignUp} options={MyTransition} />
-              <Stack.Screen
-                name="UploadedInventory"
-                component={UploadedInventory}
-                options={MyTransition}
-              />
-              <Stack.Screen name="SelectSpecies" component={SelectSpecies} options={MyTransition} />
-              <Stack.Screen name="Logs" component={Logs} options={MyTransition} />
-              <Stack.Screen name="ManageSpecies" component={ManageSpecies} option={MyTransition} />
-              <Stack.Screen name="SpecieInfo" component={SpecieInfo} option={MyTransition} />
-              <Stack.Screen name="Legals" component={Legals} options={MyTransition} />
-            </>
-          )}
-        </Stack.Navigator>
-      </NavigationContainer>
+      <AppNavigator />
     </Provider>
   );
 };
