+++ app/languages/en/static/initialScreen.json
@@ -1,4 +1,5 @@
 {
-    "updating_database":"Updating database",
-    "updating_species":"Updating species"
+    "updating_species":"Updating species",
+    "running_migration":"Running a migration",
+    "migration_description":"You recently updated your app. Please do not close the app as we are applying changes to ensure compatibility."
 }
